+++ tests/testthat/test-S3.R
@@ -69,6 +69,23 @@ test_that("vcov function works", {
     expect_named(z, c("mean", "dispersion"))
     expect_equal(z$mean, x)
     expect_equal(z$dispersion, y)
+
+    dat <- load_data("chickenpox", directory = tempdir())
+    chickenpox <- dat$chickenpox
+    population_hungary <- dat$population_hungary
+    W_hungary <- dat$W_hungary
+
+    fit2 <- dglmstarma(chickenpox, list(past_obs = 1, past_mean = 1), list(intercept = "inhomogeneous"), mean_family = vquasipoisson("identity"),
+                   dispersion_link = "log",
+                   wlist = W_hungary, 
+                   mean_covariates = list(population = population_hungary),
+                   control = list(use_fast_if_const_dispersion = TRUE))
+    z <- vcov(fit2, "dispersion")
+    expect_true(is.matrix(z))
+    expect_true(is.numeric(z))
+    expect_equal(dim(z), c(20, 20))
+    expect_true(all(eigen(z)$values > 0))
+
 })
 
 testthat::skip_on_cran()
@@ -264,9 +281,74 @@ test_that("fitted function works", {
 
 
 
+testthat::skip_on_cran()
+test_that("summary.glmstarma function works", {
+    dat <- load_data("chickenpox", directory = tempdir())
+    chickenpox <- dat$chickenpox
+    population_hungary <- dat$population_hungary
+    W_hungary <- dat$W_hungary
+
+    fit <- glmstarma(chickenpox, list(past_obs = 1), W_hungary, family = vpoisson("log"),
+                    covariates = list(population = population_hungary))
+    x <- summary(fit)
+    expect_s3_class(x, "summary.glmstarma")
+    print(x)
+
+    fit$coefficients <- NULL
+    fit$coefficients_list <- NULL
+    x <- summary(fit)
+    expect_s3_class(x, "summary.glmstarma")
+    expect_equal(length(x), 1)
+    print(x)
+
+    fit <- glmstarma(chickenpox, list(intercept = "inhomogeneous", past_obs = 1), W_hungary, family = vpoisson("log"),
+                 covariates = list(population = population_hungary))
+    x <- summary(fit)
+    print(x)
+
+    fit <- glmstarma(chickenpox, list(past_obs = cbind(c(1, 1), c(0, 1))), W_hungary, family = vpoisson("log"),
+                 covariates = list(population = population_hungary))
+    x <- summary(fit)
+    print(x)
+
+    fit <- glmstarma(chickenpox, list(past_obs = c(1, 1), past_mean = 1), W_hungary, family = vpoisson("log"),
+                  covariates = list(population = population_hungary))
+    x <- summary(fit)
+    print(x)
+
+    fit <- glmstarma(chickenpox, list(past_obs = c(1, 1)), W_hungary, family = vquasipoisson("identity"),
+                  covariates = list(population = population_hungary))
+    x <- summary(fit)
+    print(x)
+})
 
 
 
+
+testthat::skip_on_cran()
+test_that("summary.dglmstarma function works", {
+    dat <- load_data("chickenpox", directory = tempdir())
+    chickenpox <- dat$chickenpox
+    population_hungary <- dat$population_hungary
+    W_hungary <- dat$W_hungary
+
+    fit2 <- dglmstarma(chickenpox, list(past_obs = 1, past_mean = 1), list(past_obs = 1), mean_family = vquasipoisson("identity"),
+                   dispersion_link = "log",
+                   wlist = W_hungary, 
+                   mean_covariates = list(population = population_hungary))
+    x <- summary(fit2)
+    expect_s3_class(x, "summary.dglmstarma")
+    print(x)
+
+    fit2 <- dglmstarma(chickenpox, list(past_obs = 1, past_mean = 1), list(intercept = "inhomogeneous"), mean_family = vquasipoisson("identity"),
+                   dispersion_link = "log",
+                   wlist = W_hungary, 
+                   mean_covariates = list(population = population_hungary),
+                   control = list(use_fast_if_const_dispersion = TRUE))
+    x <- summary(fit2)
+    expect_s3_class(x, "summary.dglmstarma")
+    print(x)
+})