+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New Features
 - Added argument `const` to the `vgamma()` family function, which is used in the transformation of past observations when the log-link is used. Past observations are now transformed as `log(y + const)` instead of `log(y)`.
 - Added argument `dispersion_constant` to the `dglmstarma.control()` and `glmstarma_sim.control()` function. The value of this argument is passed to the `const` argument of the `vgamma()` family function when the log-link is used for the dispersion model in the `dglmstarma`-function.
+- Added `withr` to the `Suggests` field in the DESCRIPTION file. The package is now used in the `test-data.R` test.
 
 
 ## Bug Fixes
@@ -12,6 +13,7 @@
 - Fixed bug in the `variance()` and `dev.resids()` functions of the `vbinomial()` and `vquasibinomial()` family functions
 - Fixed bug in the `variance()` function of the `vnormal()` family function, which now returns the correct dimensions when the `ignore_dispersion` argument is set to `TRUE` and `mu` is not a matrix.
 - Fixed bug bug in `fitted.dglmstarma()` which prevented to remove initial time points when `drop_init = TRUE`.
+- `print.summary.glmstarma()` and `print.summary.dglmstarma()` can now handle empty models
 
 
 # glmsSTARMA 1.0.1